A significant change: scripts can now have arbitrary arguments, not just a parameter file (see ticket:12).
It is also possible to have no parameter file, but it is not currently allowed to have more than one.

Arguments that correspond to file names but are not parameter files are assumed to be input data. This is not yet finished, through, as arguably we should use a `DataStore` here, as for output data.