flex. table width: word-break
Using the word-break property in <td>
allows for word wrapping in the table
cells. Note that word wrap reportedly
only works if a width is specified,
which here is not the case.

Word-break can cause unfortunate break
points, leaving single letters in a row,
for example, reducing readability.
Breaking at underscores, additionally to
hyphens may be a better alternative. This
is explored in a subsequent commit.