Reverted list wraps in iterations
The dict api has changed: while for Python2 .iter(), .values() and
.keys() return lists, for Python3 they return views on the dict.
Hence the lib2to3.fixes.fix_dict fixer from future wraps them in a
list to be consistent. As the views are iterable as well, that
wrapping is not necessary in iterations and therefore here reverted.

There is one corner case though: in Python2 the iterable is a copy
of the dict, a list. In Python3 it is the dict itself. So should
the dict be modified during iteration, the list wrap (or any other
copy) is necessary.