Removing fragile string comparison
String comparison using ``s1 is s2`` is fragile as it might work
in some occasions (due to internal string handling of the Python
interpreter), usually it can (or should) fail.

http://stackoverflow.com/questions/1504717/

As string comparison is slow, we should consider modeling the 'default'
case with either ``None`` or a constant defined in projects.py and then
use the identy check ``is None`` or ``is OUR_CONSTANT``. This will
break the API though.