Updated GitPython requirement to 1.0.1, and fixed issue in finderepo().
An issue arose from this situation: There is a git repository MyLibs,
containing within it a list of git submodules (MyLib1, MyLib2, etc.).
Each of these is a Python module/package, so each could be used by
import MyLib1, import MyLib2, and so on. When sumatra checked the
dependencies, it would look at each of MyLib1, MyLib2, etc. and find
that their .git directories were stored as MyLibs/.git/modules/MyLib1,
for example. This meant that sumatra believed the root of each of
these modules to be the directory MyLibs/.git/modules, which is
clearly not the case.

This commit fixes the issue by basically returning the working
directory of the repository, which is either the root of the working
tree, or the .git repository if it is a bare repo. Since I don't know
what version this was introduced in, I've updated the requirement for
gitpython to 1.0.1.

It's possible I misunderstood the purpose of findrepo(), in which case
this is not a valid fix. Another potential fix might be to check
whether repo.has_separate_working_tree(), and if so, return
repo.working_tree_dir.