finding git content directly by hash
The GitWorkingCopy.content method was iterating over all commits of
the master branch before to find the commit with the specified hash.
When the commit was not on the master branch, that would fail. Instead,
it is not iterated over a branch anymore, but the commit is directly
accessed through GitPython.