Make the NTParameter.save method compatible with NeuroTools
NeuroTools' ParameterSet class does not provide an 'add_extension', as the other
parameter types defined in Sumatra do. This leads to problems as other functions
pass that keyword explicitly.
This commit adds a save method to the empty NTParameter class, which catches the
call and treats the 'add_extension' keyword, before forwarding the call to
the parameter's save method as before.