Make RecordDifference more robust
In cases where record parameters are malformed, they may not load
correctly as a dict, and then `rec.parameters.pop` raises AttributeError.
One may still want to compare the record to another however – for example,
to identify differences with a record where parameters are not malformed.
Adding a `hasattr` guard helps with this corner case, and as far as
I can tell has no downsides.