Make Django dependency optional
Django introduces many dependencies which are not backward compatible.
This is especially problematic while Sumatra still requires the
extremely old Django<2, which causes conflicts with almost any other
package depending on Django. (Not to mention “severe security” warnings
from GitHub because Django 1 is now considered insecure.)

One way to work around this is to not use the Django record store —
most likely the ShelveRecordStore. However, as long Django is listed in
the required dependencies, it causes a whole lot of dependency resolution
headaches: warnings, and in the case of conda, extremely slow
dependency resolution. This commit makes Django an optional
dependency, so that users can decide whether the benefits of the
DjangoRecordStore are worth the more complicated dependency resolution.

Changed files:
- setup.py: Make dependency optional
- sumatra/projects.py: Fallback if django cannot be imported