Internalise code for tag management
Since the version of django-tagging currently available on PyPI doesn't work with Django 5,
and PyPI doesn't allow dependencies that aren't in PyPI, I have moved parts
of the django-tagging code to within the django_store module,
as allowed by the BSD license of django-tagging.

The license notice from django-tagging has been appended to the LICENSE file of Sumatra.